Refatoração dos testes.
close #21
